+++ druid-shell/src/backend/windows/window.rs
@@ -161,6 +161,7 @@ enum DeferredOp {
     SetSize(Size),
     SetResizable(bool),
     SetWindowState(window::WindowState),
+    ShowWindow(bool),
     ReleaseMouseCapture,
 }
 
@@ -618,6 +619,17 @@ impl MyWndProc {
                         ShowWindow(hwnd, show);
                     }
                 }
+                DeferredOp::ShowWindow(should_show) => {
+                    let show = if should_show { SW_SHOW } else { SW_HIDE };
+
+                    unsafe {
+                        ShowWindow(hwnd, show);
+
+                        if should_show {
+                            SetForegroundWindow(hwnd);
+                        }
+                    }
+                }
                 DeferredOp::SaveAs(options, token) => {
                     let info = unsafe {
                         get_file_dialog_path(hwnd, FileDialogType::Save, options)
@@ -758,6 +770,9 @@ impl WndProc for MyWndProc {
                 Some(0)
             }
             WM_ACTIVATE => {
+                // Check if the low-order word is not 0
+                // If it were 0, then that means we are being deactivated, in which case we do nothing.
+                // If it is != 0, then we have been activated.
                 if LOWORD(wparam as u32) as u32 != 0 {
                     unsafe {
                         if !self.has_titlebar() && !self.is_transparent() {
@@ -1850,6 +1865,7 @@ impl WindowHandle {
         }
     }
 
+    /// Closes the window.
     pub fn close(&self) {
         if let Some(w) = self.state.upgrade() {
             let hwnd = w.hwnd.get();
@@ -1859,10 +1875,14 @@ impl WindowHandle {
         }
     }
 
+    /// Hides the window.
+    pub fn hide(&self) {
+        self.defer(DeferredOp::ShowWindow(false));
+    }
+
     /// Bring this window to the front of the window stack and give it focus.
     pub fn bring_to_front_and_focus(&self) {
-        //FIXME: implementation goes here
-        warn!("bring_to_front_and_focus not yet implemented on windows");
+        self.defer(DeferredOp::ShowWindow(true));
     }
 
     pub fn request_anim_frame(&self) {